+++ sponge/api7/src/main/java/com/cadiducho/cservidoresmc/SpongePlugin.java
@@ -31,7 +31,7 @@ import java.util.List;
 @Plugin(id = "cservidoresmc", name = "40ServidoresMC", version = SpongePlugin.PLUGIN_VERSION)
 public class SpongePlugin implements CSPlugin {
 
-    public static final String PLUGIN_VERSION = "3.0.1";
+    public static final String PLUGIN_VERSION = "3.0.2";
     @Inject private Logger logger;
     @Inject private Game game;